Consolidate all Git repo functions
