+++ vulnfeeds/git/repository_test.go
@@ -1,6 +1,7 @@
 package git
 
 import (
+	"reflect"
 	"testing"
 
 	"github.com/go-test/deep"
@@ -131,6 +132,67 @@ func TestRepoTags(t *testing.T) {
 		}
 		if tc.cache != nil && !(cache_after > cache_before) {
 			t.Errorf("test %q: RepoTags(%q) incorrect cache behaviour: size before: %d size after: %d", tc.description, tc.inputRepoURL, cache_before, cache_after)
+		}
+	}
+}
+
+func TestValidRepo(t *testing.T) {
+	tests := []struct {
+		description    string
+		repoURL        string
+		expectedResult interface{}
+		expectedOk     bool
+	}{
+		{
+			description:    "Valid repository",
+			repoURL:        "https://github.com/torvalds/linux",
+			expectedResult: true,
+			expectedOk:     true,
+		},
+		{
+			description:    "Invalid repository",
+			repoURL:        "https://github.com/andrewpollock/mybogusrepo",
+			expectedResult: false,
+			expectedOk:     true,
+		},
+	}
+	for _, tc := range tests {
+		got, err := ValidRepo(tc.repoURL)
+		if err != nil && tc.expectedOk {
+			t.Errorf("test %q: ValidRepo(%q) unexpectedly failed: %#v", tc.description, tc.repoURL, err)
+		}
+		if !reflect.DeepEqual(got, tc.expectedResult) {
+			t.Errorf("test %q: ValidRepo(%q) was incorrect, got: %#v, expected: %#v", tc.description, tc.repoURL, got, tc.expectedResult)
+		}
+	}
+}
+func TestValidRepo(t *testing.T) {
+	tests := []struct {
+		description    string
+		repoURL        string
+		expectedResult interface{}
+		expectedOk     bool
+	}{
+		{
+			description:    "Valid repository",
+			repoURL:        "https://github.com/torvalds/linux",
+			expectedResult: true,
+			expectedOk:     true,
+		},
+		{
+			description:    "Invalid repository",
+			repoURL:        "https://github.com/andrewpollock/mybogusrepo",
+			expectedResult: false,
+			expectedOk:     true,
+		},
+	}
+	for _, tc := range tests {
+		got, err := ValidRepo(tc.repoURL)
+		if err != nil && tc.expectedOk {
+			t.Errorf("test %q: ValidRepo(%q) unexpectedly failed: %#v", tc.description, tc.repoURL, err)
+		}
+		if !reflect.DeepEqual(got, tc.expectedResult) {
+			t.Errorf("test %q: ValidRepo(%q) was incorrect, got: %#v, expected: %#v", tc.description, tc.repoURL, got, tc.expectedResult)
 		}
 	}
 }
